backend/requirements.txt
fastapi
uvicorn
langchain_community
langchain
langchain_openai
python-dotenv
python-multipart
chromadb
PyPDF2
python-docx